% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_tdr.R
\name{calibrate_tdr}
\alias{calibrate_tdr}
\title{Calibrate CEFAS data}
\usage{
calibrate_tdr(tdr, deployid, rate = 0.1, surface_thr = 0.1,
  depth_thr = 0.2, dur_thr = 0.5)
}
\arguments{
\item{tdr}{data.frame. Use \code{\link{read_cefas}} to parse CEFAS file.}

\item{rate}{numeric. Sampling rate for FastLog events.}

\item{surface_thr}{numeric. Threshold for surface noise.}

\item{depth_thr}{numeric. Minimum depth to qualify as a dive.}

\item{dur_thr}{numeric. Minimum duration to qualify as a dive.}

\item{id}{character. Deployment identifier.}
}
\value{
data.frame with columns
\itemize{
  \item{\code{Row} (Row number in original CEFAS file)}
  \item{\code{UTC} (Datetime of record in UTC timezone)}
  \item{\code{Pressure}}
  \item{\code{EventID} (FastLog identifier)}
  \item{\code{CalibPressure} (Pressure after surface offset)}
  \item{\code{Surface} (Surface offset)}
  \item{\code{DivePhase} (From diveMove)}
  \item{\code{duration}}
  \item{\code{depth}}
  \item{\code{DiveID}}
  \item{\code{DeployID}}
}
}
\description{
\code{calibrate_tdr} is a wrapper for
\code{\link{diveMove::calibrateDepth}}. CEFAS tags record depth in bursts,
so this function uses a split-apply-combine approach to calibrate each
FastLog individually. Tags have an error of +/-1m, so for shallow diving
birds the error may be greater than the depth of the dive. To avoid losing
these dives, the "surface" is assumed to be at the shallowest point in each
FastLog, not at 0. For tags with a 1m threshold, this may bias dive depths
shallower.
}
\examples{
# Uses tidyverse functions
metadata <- system.file('extdata',
  'MOC2015PFSHmetadata.csv',
  package = 'pfsh.dive') \%>\%
  read_csv
example_tdr <- system.file('extdata',
  paste0('metadata$TDR_filename[1]', '.CSV'),
  package = 'pfsh.dive')
tdr <- read_cefas(example_tdr,
  metadata$Deployed[1],
  metadata$Recovered[1])
calib_tdr <- calibrate_tdr(tdr, metadata$DeployID[1])

}
