% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dive.R
\name{plot_dive}
\alias{plot_dive}
\title{Create a dive plot}
\usage{
plot_dive(calib_tdr, diveid, surface_thr = 0.1)
}
\arguments{
\item{calib_tdr}{data.frame. Use \code{\link{calibrate_tdr}}.}

\item{diveid}{character. Deployment identifier.}

\item{surface_thr}{numeric. Threshold for surface noise.}
}
\value{
gtable with two plots.
}
\description{
\code{plot_dive} creates a plot of a single dive in two panels. The left
panel zooms in on the dive itself and the right panel is an overview of the
entire event. In the left panel, blue points are within the dive and red
points are the surrounding records for context. The right panel has two
curves. The red one shows the original, pre-calibration points. The green
one is the calibrated dive with the surface offset applied. The shaded
region delineates the dive itself. In both panels, the dashed blue line
indicates the surface noise threshold.
}
\examples{
# Uses tidyverse functions
metadata <- system.file('extdata',
  'MOC2015PFSHmetadata.csv',
  package = 'pfsh.dive') \%>\%
  read_csv
example_tdr <- system.file('extdata',
  paste0('metadata$TDR_filename[1]', '.CSV'),
  package = 'pfsh.dive')
tdr <- read_cefas(example_tdr,
  metadata$Deployed[1],
  metadata$Recovered[1])
calib_tdr <- calibrate_tdr(tdr, metadata$DeployID[1])
plot_dive(calib_tdr, 4)

}
