% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cefas.R
\name{read_cefas}
\alias{read_cefas}
\title{Read a CEFAS file}
\usage{
read_cefas(cefas_file, deployed, recovered)
}
\arguments{
\item{cefas_file}{character. Path to CEFAS file.}

\item{deployed}{POSIXct. Datetime of deployment.}

\item{recovered}{POSIXct. Datetime of recovery}
}
\value{
data.frame with columns
\itemize{
  \item{\code{Row} (Row number in original CEFAS file)}
  \item{\code{UTC} (Datetime of record in UTC timezone)}
  \item{\code{Pressure}}
  \item{\code{EventID} (FastLog identifier)}
}
}
\description{
Downloading data from CEFAS tags yields a non-tabular CSV. \code{read_cefas}
reads a CEFAS file and extracts the depth records in a tidy data.frame.
}
\examples{
# Load metadata
metadata_path <- system.file('extdata',
                             'MOC2015PFSHmetadata.csv',
                             package = 'pfsh.dive')
metadata <- readr::read_csv(metadata_path)

# Read CEFAS output
tdr_path_ <- system.file('extdata',
                         paste0(metadata$TDR_filename[1], '.CSV'),
                         package = 'pfsh.dive')
tdr <- read_cefas(example_tdr,
                  metadata$Deployed[1],
                  metadata$Recovered[1])

}
